# Optimize throttler for cases where the breaker is already present.