<!--
Request Prow to automatically lint any go code in this PR:

/lint
-->

## Proposed Changes

**Note:** This might very well be premature optimization but I thought the fix is so small and not ultra complex so it might make sense to take it in.

`getOrCreateBreaker` checks whether a breaker is already present. If so, it returns it. If not, it creates a new one. To do so though, it needs to acquire a write-lock. Always acquiring that write-lock results in a lot of contention and other requests, even if they go to an entirely different revision, are excluded from that path.

I investigated the usage of a concurrent map for this use-case and while that might still be a valuable thing to do, it comes with a few caveats. Most notably, all concurrent maps I could find have no method to lazily compute a value if it's not yet set. Since we atomically need to get-or-set (we only want to create one instance of the breaker) we'd need to generate the new breaker on every call. Not nice.

Instead, this relies on an optimistic approach. It acquires a read-lock first and tries to get a breaker. If no breaker is present, we get a write-lock and create a new breaker after checking, if we haven't been raced. That reduces the contention in the happy case (which is the vast majority of proxying requests) to just read-locks.

## Benchmark Results

### Old

```
go test -benchmem -run=^$  knative.dev/serving/pkg/activator -bench "^(BenchmarkThrottler)$"
goos: darwin
goarch: amd64
pkg: knative.dev/serving/pkg/activator
BenchmarkThrottler/1-12  	10000000	       156 ns/op	       0 B/op	       0 allocs/op
BenchmarkThrottler/10-12 	10000000	       162 ns/op	       0 B/op	       0 allocs/op
BenchmarkThrottler/100-12         	10000000	       220 ns/op	       0 B/op	       0 allocs/op
BenchmarkThrottler/1000-12        	 2000000	       669 ns/op	       2 B/op	       0 allocs/op
BenchmarkThrottler/10000-12       	 5000000	       743 ns/op	       1 B/op	       0 allocs/op
PASS
```

### New

```
go test -benchmem -run=^$  knative.dev/serving/pkg/activator -bench "^(BenchmarkThrottler)$"
goos: darwin
goarch: amd64
pkg: knative.dev/serving/pkg/activator
BenchmarkThrottler/1-12  	30000000	        45.8 ns/op	       0 B/op	       0 allocs/op
BenchmarkThrottler/10-12 	30000000	        45.5 ns/op	       0 B/op	       0 allocs/op
BenchmarkThrottler/100-12         	30000000	        45.2 ns/op	       0 B/op	       0 allocs/op
BenchmarkThrottler/1000-12        	30000000	        45.4 ns/op	       0 B/op	       0 allocs/op
BenchmarkThrottler/10000-12       	30000000	        46.3 ns/op	       0 B/op	       0 allocs/op
PASS
```

This shows nice linear scalability of the `Try` method regardless of the number of goroutines (i.e. requests) hitting it under load.

**Release Note**

<!-- Enter your extended release note in the below block. If the PR requires
additional action from users switching to the new release, include the string
"action required". If no release note is required, write "NONE". -->

```release-note
NONE
```

/assign @vagababov